Added script to scrape the NHS Catalogue
To create a MVP of showing metadata
from an API Catalogue

This is to scrape the NHS API Catalogue:
https://digital.nhs.uk/developer/api-catalogue

We've made some assumptions about the page:

Some pages have a slightly different structure
with an svg in the overview section. In this case
we get the div after the svg and return that as
the description.

The urls are gathered based off an existing
structure of /developer/api-catalogue/ which
if changes will result in a failed run